--- env.bat
@@ -1,3 +1,0 @@
-CALL conda create --name test
-conda install pandas
-conda install numpy
