--- run.bat
@@ -1,3 +1,0 @@
-cd "E:\single-input-gtd\todo"
-CALL conda activate test
-python UI.py
