+++ time_record/2022-07-31.csv
@@ -1,3 +1,4 @@
 name,subname,actuall_start_time,actuall_end_time,actlength,actshift,reason
 测试2.0,测试用,2022-07-31 00:04:45.702431,2022-07-31 00:04:53.574120,0,0,finished
 测试子项目,测试中，哈哈,2022-07-31 00:25:10.621223,2022-07-31 00:25:17.325789,0,0,finished
+线性代数,约尔丹和一些复数化,2022-07-31 14:20:11.979371,2022-07-31 15:35:19.538605,0,0,finished
