--- user_input.bat
@@ -1,3 +1,0 @@
-cd "E:\single-input-gtd"
-CALL conda activate test
-python user_input.py
